requirements.txt
requests
google-genai
dotenv
opencv-python
Pillow